+++ components/nfc/mfkey32/screen/src/main/java/com/flipperdevices/nfc/mfkey32/screen/composable/progressbar/ComposableMfKey32NotFound.kt
@@ -5,9 +5,10 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
@@ -16,10 +17,13 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.core.markdown.annotatedStringFromMarkdown
+import com.flipperdevices.core.preference.pb.HardwareColor
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.nfc.mfkey32.screen.R
+import com.flipperdevices.nfc.mfkey32.screen.viewmodel.FlipperColorViewModel
+import tangle.viewmodel.compose.tangleViewModel
 
 @Composable
 fun ComposableMfKey32NotFound() = Column(
@@ -33,11 +37,15 @@ fun ComposableMfKey32NotFound() = Column(
         color = LocalPallet.current.text100,
         textAlign = TextAlign.Center
     )
+    val flipperColorViewModel = tangleViewModel<FlipperColorViewModel>()
+    val flipperColor by flipperColorViewModel.getFlipperColor().collectAsState()
     Image(
         painter = painterResource(
-            if (MaterialTheme.colors.isLight) {
-                R.drawable.pic_flipper_nfc_detect_reader_white
-            } else R.drawable.pic_flipper_nfc_detect_reader_black
+            when (flipperColor) {
+                HardwareColor.WHITE,
+                HardwareColor.UNRECOGNIZED -> R.drawable.pic_flipper_nfc_detect_reader_white
+                HardwareColor.BLACK -> R.drawable.pic_flipper_nfc_detect_reader_black
+            }
         ),
         contentDescription = stringResource(R.string.mfkey32_not_found_title)
     )